refactor with single command dockerfile
