+++ cmd/root.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"github.com/la-plas-growth/go-docker-ai/cmd/command/dockerfile"
-	"github.com/la-plas-growth/go-docker-ai/cmd/command/lint"
 	"github.com/spf13/cobra"
 )
 
@@ -11,9 +10,9 @@ func NewRootCommand() *cobra.Command {
 		Use:   "tool",
 		Short: "Tool is a CLI application for lint.",
 	}
-	rootCmd.AddCommand(lint.NewBaseCommand())
-	rootCmd.AddCommand(dockerfile.CreateDockerfileCommand())
-
+	//
+	rootCmd.AddCommand(dockerfile.NewBaseCommand())
+	//
 	return rootCmd
 }
