+++ examples/dockerfile.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 go build .
-./go-docker-ai dockerfile -t golang   
+#
+./go-docker-ai dockerfile create -t golang -p <path>
+#
+./go-docker-ai dockerfile lint -f <path>
+#
