--- examples/lint.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-go build .
-./go-docker-ai dockerlint -f Dockerfile
